Reorganize files for maintainability and to simplify imports.
