+++ .vscode/settings.json
@@ -13,10 +13,14 @@
         "argcount",
         "co_argcount",
         "conditionfn",
+        "destroyfn",
+        "fillvalue",
         "fpipeline",
+        "heml",
         "kvars",
         "kwvars",
         "lcond",
+        "mkrow",
         "ndict",
         "nval",
         "stepfn",
